--- money/dollar.py
@@ -1,18 +1,0 @@
-
-class Dollar(object):
-    """docstring for Dollar"""
-    def __init__(self, amount):
-        super(Dollar, self).__init__()
-        self._amount = amount
-
-    def times(self, multiplier):
-        return Dollar(self._amount * multiplier)
-
-    def __eq__(self, other):
-        if isinstance(other, self.__class__):
-            return self.__dict__ == other.__dict__
-        else:
-            return False
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
